Move genmlir to genbenchmark.py
